Move segment_anything_model.py to ai directory
